[bugfix] Consistently indent declarations.
To prevent losing tabs from indentation in declarations, FreeBSD indent's r125624 added code for the most common case when it's an identifier that is indented, but didn't do anything with the original code that did the same for any other cases. The other cases are: lparens (function pointer declaration), asterisks (pointer declaration), stray semicolons, and commas leading identifiers instead of trailing them.

Use the code added in r125624 (and improved in later commits) to write a new function indent_declaration() and use it in all places that meant to indent declarations. In order to indent only once per line, reuse existing ps.dumped_decl_indent variable that was only used when formatting for troff (-troff) until now.