[bugfix] Removed whitespace shouldn't be considered in column calculations.
This piece of code removed tabs and space characters from after colons that follow labels by decrementing the e_lab (end of label) "pointer" which is later used to calculate the width of the string that fprintf() puts into "output". But pad_output() gets the length from the actual string, so it miscalculated what the current column is.

Fixed by putting a string terminator at the e_lab "pointer".