[cleanup] Remove dead code relating to unix-style comments.
The original indent(1) described unix-style comments as similar to box comments, except the first non-blank character on each line is lined up with the '*' of the "/*" which appears on a line by itself.

This code has always been a hack as it tried to guess whether a comment is a box comment or a unix-style comment and treat each comment according to its own style. It's been turned off for ages and -sc/-nsc make it even less relevant.