[bugfix] Fix wrapping of some lines in comments.
After a blank line was printed (to separate paragraphs in comments), the next line was sometimes wrapped to the column at which the previous non-empty line ended. The fix is to reset the last blank pointer (last_bl) on newline. Also use the NULL macro instead of constant 0 for clarity, because it's a pointer.