[feature] Support binary integer literals.
This was done by Romain for PR123553. I initially thought that it would break code like this:
#define b00101010 -1
if (0 b00101010)
...

by joining 0 and b00101010 together. However, the real problem with that patch was that once it saw a 0, it assumed that the number was base 2, 8 or 16, ignoring base 10 floating point numbers like 0.1. I fixed that.

I didn't copy the diagnostic part of the original patch as it seems out of scope of implementing binary integer literals.