[bugfix] Improve handling of comments and newlines between "if (...)" or "while (...)" and "else" or "{".
Don't flush newlines - there can be multiple of them and they can happen before a token that isn't else or {. Instead, always store them in save_com.
Don't dump the buffer's contents on newline assuming that there is only one comment before else or {.
Try to avoid producing surplus newlines, especially before else when -ce is on.
When -bl is on, don't treat { as a comment (was implemented by falling through "case lbrace:" to "case comment:").