[bugfix] Improve predictability of lexi().
The lexi() function reads input stream and categorizes the next token. indent(1) will sometimes buffer up a sequence of tokens in order rearrange them. That is needed for properly cuddling else or putting braces correctly according to the chosen style (KNF vs Allman) when comments are around. The loop that buffers tokens up uses lexi() to decide if it's time to stop buffering. Then the temporary buffer is used to feed lexi() the same tokens again, this time for normal processing.

The problem is that lexi() apart from recognizing the token, can change a lot of information about the current state, for example ps.last_nl, ps.keyword, buf_ptr. It also abandons leading whitespace, which is needed mainly for comment-related considerations. So the call to lexi() while tokens are buffered up and categorized can change the state before they're read again for normal processing which may easily result in changing interpretation of the current state and lead to incorrect output.

To work around the problems:
1) copy the whitespace into the save_com buffer so that it will be read again when processed
2) trick lexi() into modifying a temporary copy of the parser state instead of the original.